Added client-side main-category swapping to item editing to close issue #19. There are now two drop-down menus, one for the main category and one for the sub. When you swap the main category, an ajax call goes off to repopulate the sub-categories.
Altered the GameItemViewModel to reflect the addition of a separate listing, as well as the ViewModel population in the ItemsController. Also added another data request action to the ItemsController for getting main categories.

The script accounts for new game items without category, and will automatically set the correct category selections when editing a previous entry.